Solutions - Build a Calculator
Build a calculator (add, subract, multiply and divide). Ask the user for 1st number , 2nd number and operation. Print out the inputs, operation and answer. Ask the user if will continue in calculation using the answer or start a new calculation. If the user wants to start a new calculation, clear the terminal .